requirements.txt
Flask==3.0.0
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-Limiter==3.5.0
pymongo==4.6.1
dnspython==2.4.2
cryptography==42.0.2
scikit-learn==1.4.0
numpy==1.26.4
httpx==0.26.0
pypdf==4.0.1
python-docx==1.1.0